requirements.txt
Flask==2.2.2
flask_sqlalchemy==3.0.2
python-dotenv==0.21.0
twilio==7.16.1
